Add .idea configuration files and contents of 401.html and allocate.html
Added .idea module configuration files for the project. Also added the contents of files 401.html and allocate.html. The .idea files help in managing project settings and are automatically generated when a project is opened in IDEA. The html files were added for the functionality of the website, one is for the 401 error page and the other is for allocating resources on the website.