Add 'hide' class to page_loader divs for smoother rendering
The 'hide' class has been added to the page_loader divs across multiple HTML files. This change is in accordance with a design decision to prevent the premature display of the loading screen before the whole page is completely loaded. This enhancement aims to improve the user experience by avoiding the rendered display of a partially loaded page.