Change visibility of 'body_content' divs in settings/member.html and sign-in.html
The 'visible' class has been removed from 'body_content' divs in settings/member.html and sign-in.html. This change is aimed at improving initial rendering of these pages by not forcing the visibility of the body content during page load. This should result in smoother rendering and better presentation of page content to the user.