Fix visibility issue for 'sign_up_wrp' and 'sign_in_wrp' in sign-up and sign-in pages
Previously, the divs 'sign_up_wrp' and 'sign_in_wrp' on the sign-up and sign-in pages respectively had a forced opacity of 0. This seemed to hamper the visibility of important div elements upon initial page loading leading to an undesirable user experience. In this commit, the forced opacity has been removed which should ensure smoother rendering and correct visibility of these elements right from the start.