Update API Reference link in developers.html
Changed the URL of the "API Reference" link in the developers.html file. The URL was previously pointed to an external domain "illusory.io/docs" which is now redirected to a local document "/docs/api-reference/introduction". This was done to ensure that users are directed to the most current and relevant documentation within the same site by default.