Correct URL paths in member discount links
The URL paths to member discount settings in the new, allocate, and dashboard pages have been updated. Also, on the dashboard page, the connection instructions link has been directed to the appropriate documentation page. These changes ensure that users are correctly directed to the intended pages upon clicking the links.