Remove public/start.html file
The HTML file for the start page (public/start.html) was removed completely. This change might be part of a restructuring effort, perhaps due to changes in the website navigation flow or a switch to a single-page application model. Remember to ensure all relevant routes and references are updated accordingly.