Update CSS and JS source paths across all HTML files
The source paths for all CSS and JS files across all HTML pages were changed. The URLs now point to the '/app' directory in the '/www.illusory.io' domain, instead of the previous individual directories. This change has been done to maintain consistency and to allow all files to be retrieved from the same place, simplifying management and potential debugging.